${2AM11:17 AMe11:17 AMi17e}

+++ MazeGameCode/Mazey-4-gamerun.py
@@ -61,7 +61,7 @@ class Player(pygame.sprite.Sprite):
                 self.rect.left = wall.rect.right
         
         if keys[pygame.K_UP]:
-            self.rect.y -=(PLAYER_SPEED * delta)
+            self.rect.y -= int(PLAYER_SPEED * delta)
         if keys[pygame.K_DOWN]:
             self.rect.y += int(PLAYER_SPEED * delta)
